% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_death_row.R
\name{scrape_death_row}
\alias{scrape_death_row}
\title{Scrape Texas Department of Criminal Justice website}
\usage{
scrape_death_row()
}
\value{
ExecutedOffenders table of executed offenders
}
\description{
Scrape Texas Department of Criminal Justice website
}
\examples{
# not run
# ExecutedOffenders <- scrape_death_row()
}
